Changes the scoping of states
Previously states were scoped according to projects that is if two
projects are present in same state, both will have their own instance
of that state in terms of verticals and users in those verticals but
now the state is a constant and each project that picks up that state
will see it as it is. So all projects in one state will see it exactly
the same.